Generalize SandboxDemoMessage -> SandboxMessage
