--- frontend/components/Sandbox/SandboxDemoMessage/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SandboxDemoMessage";
