+++ frontend/components/Sandbox/SandboxMessage/_styles.scss
@@ -1,15 +1,15 @@
-.sandbox-demo-message {
+.sandbox-message {
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: flex-start;
+  gap: $pad-small;
   width: 350px;
-  margin: auto;
+  margin: $pad-xxlarge auto;
 
   &__message {
     font-size: $small;
     font-weight: $bold;
-    margin: 0 0 $pad-large;
   }
 
   &__link-message {
