+++ frontend/pages/admin/UserManagementPage/UserManagementPage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { InjectedRouter } from "react-router";
 import SandboxGate from "components/Sandbox/SandboxGate";
-import SandboxDemoMessage from "components/Sandbox/SandboxDemoMessage";
+import SandboxMessage from "components/Sandbox/SandboxMessage";
 import UsersTable from "./components/UsersTable";
 
 const baseClass = "user-management";
@@ -19,10 +19,10 @@ const UserManagementPage = ({ router }: IUserManagementProps): JSX.Element => {
       </p>
       <SandboxGate
         fallbackComponent={() => (
-          <SandboxDemoMessage
+          <SandboxMessage
             message="User management is only available in self-managed Fleet"
             utmSource="fleet-ui-users-page"
-            className={`${baseClass}__sandbox-demo-message`}
+            className={`${baseClass}__sandbox-message`}
           />
         )}
       >
